Refactor tests.utils module
